+++ sbin/restart_server.sh
@@ -1,0 +1,11 @@
+#!/bin/bash -e
+
+### Restart web server
+if [ -f /etc/redhat-release ]
+then
+    service httpd restart
+else
+    service apache2 restart
+fi
+
+exit 0