+++ CHANGES.md
@@ -1,3 +1,11 @@
+# 1.1.0 (26 April 2017)
+
+* [+] Added PageSpeed Apache Module by Google - Complete integration with installation and configuration
+* [+] Improved PageSpeed Apache status check
+* [+] Added a note if domain name does not resolve to the Plesk server
+* [+] Code optimizations 
+* [-] Fixed IP resolving issue when server is behind a proxy, e.g. Cloudflare or AWS server
+
 # 1.0.1 (21 February 2017)
 
 * [+] Added links to blog articles about PageSpeed Insights to the Description and Readme files